+++ src/Soluble/Japha/Bridge/Driver/Pjb62/Client.php
@@ -443,14 +443,6 @@ class Client
     }
 
     /**
-     * @return ParserString
-     */
-    public function createParserString()
-    {
-        return new ParserString();
-    }
-
-    /**
      * @param mixed $arg
      */
     protected function writeArg($arg)